add ability to use other templates for file or subdirectory of files
For a file, specify the relative path to the desired template.

   ```elixir
   <%
     title = "Special post"
     template = "_special_template.html.eex"
   %>
   ```

For an entire subdirectory of files (like a file of blog posts in
`posts`), you can create a `_template.html.eex` file in that
directory. When the file is compiled, it will automatically look in
its current directory for a `_template.html.eex` file. It will use
that if it exists. If it does not exist, it will search the parent
directory and so on until it reaches `src_dir/_template.html.eex`.