feat: extract starting up of Vox.Builder
When the watcher has been built, we'll need to recompile the
site. Before this, we'd also try to start the Genserver again which
gives us an error that it is already started. This splits out those
tasks (starting the server and building the site) so that we can call
them separately as needed.