feat: empty builder state before building
I found that when the dev server is running for a while, it gets
really slow. My guess right now (without too much digging) is that the
state gets so large because it never empties and every file change
adds all the files back into the state again. This empties the state
before building the site.

I'm hoping that this will keep the speed of building up.

Once I'm confident about the build process, I can also eliminate
storing a lot of the intermediate steps in the build process.